requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
coverage==7.6.1
cryptography==43.0.1
diff-match-patch==20230430
dj-rest-auth==6.0.0
Django==4.2.16
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-import-export==4.1.1
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
numpy==2.1.1
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
protobuf==4.25.4
#psycopg2-binary==2.9.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
celery[redis]
psycopg2-binary==2.9.3
sentry-sdk
